rewrite xp command to ts
